SmartHire/backend2/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
mediapipe==0.10.14
deepface==0.0.93
numpy==1.26.4
Pillow==10.3.0
aiofiles==23.2.1
pydantic==2.7.1
tf-keras==2.16.0
